+++ ezrules/frontend/src/app/services/rule.service.ts
@@ -36,6 +36,7 @@ export interface RulesResponse {
 
 export interface VerifyRuleResponse {
   params: string[];
+  warnings: string[];
 }
 
 export interface TestRuleResponse {